server run with nodemon in "dev" and test in watch mode
